Moved saveLocalCart function to its own file, imported to AllProducts and SingleProduct
