+++ client/components/AllProducts.js
@@ -1,6 +1,8 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { fetchProducts } from "../store/products";
+import saveLocalCart from "../saveLocalCart";
+
 /**
  * COMPONENT
  */
@@ -15,20 +17,6 @@ export const AllProducts = (props) => {
     dispatch(fetchProducts());
   }, []);
 
-  const saveLocalCart = (item) => {
-    let cart;
-    if (localStorage.getItem("cart") === null) {
-      //If no local storage present, create a cart
-      cart = [];
-    } else {
-      //If there is, parse the local storage string, and assign cart to parsed string (which is now the cart array)
-      cart = JSON.parse(localStorage.getItem("cart"));
-    }
-    //Push the item onto the server (our computers) cart array
-    cart.push(item);
-    //Update the cart on the web's local storage
-    localStorage.setItem("cart", JSON.stringify(cart));
-  };
   return (
     <div>
       <h3>Shop Emojis!</h3>
