+++ client/components/SingleProduct.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { fetchProduct } from "../store/singleProduct";
+import saveLocalCart from "../saveLocalCart";
 
 /**
  * COMPONENT -- translating this from AllProducts
@@ -15,21 +16,6 @@ const SingleProduct = (props) => {
     console.log("Props-->", props);
   }, []);
 
-  // should we consolidate this SaveLocalCart function somewhere?
-  const saveLocalCart = (item) => {
-    let cart;
-    if (localStorage.getItem("cart") === null) {
-      //If no local storage present, create a cart
-      cart = [];
-    } else {
-      //If there is, parse the local storage string, and assign cart to parsed string (which is now the cart array)
-      cart = JSON.parse(localStorage.getItem("cart"));
-    }
-    //Push the item onto the server (our computers) cart array
-    cart.push(item);
-    //Update the cart on the web's local storage
-    localStorage.setItem("cart", JSON.stringify(cart));
-  };
 
   const caller = () => {
     saveLocalCart(product);
@@ -54,14 +40,4 @@ const SingleProduct = (props) => {
   );
 };
 
-/**
- * CONTAINER
- */
-// const mapState = (state) => {
-//   return {
-//     username: state.auth.username,
-//   };
-// };
-
-// export default connect(mapState)(SingleProduct);
 export default SingleProduct;
